requirements.txt
dj-database-url==0.4.1
Django==1.10
django-crispy-forms==1.6.0
djangorestframework==3.4.6
drfdocs==0.0.11
gunicorn==19.6.0
hashids==1.1.0
psycopg2==2.6.2
whitenoise==3.2.1
